==> analysis_files_revision/01_load_new_samples.Rmd <==
---
title: "R Notebook"
output:
  html_notebook: default
  html_document:
    df_print: paged
  pdf_document: default
---

# Within-host diversity and minor variant analyses in samples from Houston Methodist Hospital, 2021

## Sample characteristics and inclusion criteria 


```{r}
source("./scripts/startup.R")

##### Load libraries and define recurring functions / variables
#all mcov samples we have ever received - reformat sample names to be consistent
mcov_samples_all <- fread("full_lineage_report_20220507.tsv", data.table=F) %>% 
  mutate(MCoVNumber=mcov_reformat(taxon)) %>% 
  filter(startsWith(MCoVNumber, "MCoV")) %>% 
  filter(MCoVNumber!="MCoV30904") #remove one sample with inconsistently formatted duplicates

#run and date info; keep only earliest sample from the same patient
mcov_info <- fread("sample_date_and_run.csv", data.table=F) %>% 
  mutate(COLLECTION_DT=as.Date(COLLECTION_DT, "%m/%d/%y"),
         MCoVNumber=str_remove(mcov_id, "-")) %>%
  arrange(COLLECTION_DT) %>% filter(!duplicated(PatientID))

#samples we received that we're interested in (December 2020-November 2021)
mcov_samples_1 <- mcov_samples_all %>% filter(!taxon %in% dup65.66) %>% 
  filter(MCoVNumber %in% mcov_info$MCoVNumber) %>% left_join(mcov_info)

#all samples sequenced in each run (including those from outside of this study period) -- sometimes relevant for run QC purposes
runs_all_samples <- fread("run_samples.csv", data.table = F) %>% 
  mutate(MCoVNumber=str_remove(`Sample ID`,"-"), run=Run) %>% 
  select(run, MCoVNumber)

#summary stats on coverage of each sample; drop the run 65 duplicates and join coverage info to main dataset
d1 = fread('coverage_levels_20220507.csv', data.table = F) %>% 
  filter(duplicated(samplename)) %>% pull(samplename)
d2 = fread('coverage_levels_20220507.csv', data.table = F) %>% 
  filter(samplename %in% d1) %>% filter(!duplicated(samplename)) %>% 
  pull(filename)
coverage_levels <- fread('coverage_levels_20220507.csv', data.table = F) %>% 
  filter(!filename %in% d2) %>% filter(samplename!="MCoV30904") %>% select(-1)
```

```{r}
# join dataframes
mcov_samples<-mcov_samples_1 %>% 
  select(MCoVNumber, lineage, scorpio_call, qc_status, 
         COLLECTION_DT, INSTRUMENT, INSTRUMENT_RESULT, run=run_group) %>% 
  left_join(coverage_levels, by=c("MCoVNumber"="samplename"))

mcov_samples_with_ct<-mcov_samples %>% filter(INSTRUMENT_RESULT<50) %>% 
  mutate(CT=INSTRUMENT_RESULT)
#what's the relationship between CT value and read coverage?
coverage_all<-mcov_samples_with_ct %>% ggplot(aes(x=CT, y=median_coverage)) + 
  geom_point(alpha=0.07) + theme_bw() + xlab("Ct value") + ylab("Sample median depth")
#how does coverage in high-CT samples compare with the rest of them?
coverage_ct_cat<-mcov_samples_with_ct %>% mutate(sample_ct=if_else(CT>=40, "CT>=40", "CT<40")) %>% 
  ggplot(aes(x=sample_ct, y=median_coverage)) + 
  geom_point(alpha=0.2, position=position_jitter(width=0.25)) + 
  geom_boxplot(color="red", alpha=0) + theme_bw() + 
  xlab("Ct value") + ylab("Sample median depth")
```


## Run-level QC 

```{r}
#are there any runs where high-CT samples have unusually high coverage? treat CT>40 samples as negative controls and eliminate runs where their coverage is not different from those of the rest of the samples
test_group<-mcov_samples_with_ct %>% mutate(is_neg=if_else(CT>=40,1,0)) %>% 
  group_by(run) %>% mutate(n_negs=sum(is_neg)) %>% filter(n_negs>=3) %>% 
  ungroup() %>% mutate(sampletype=if_else(is_neg==1, "negctrl","sample"))
runs_to_drop1 = test_group %>% group_by(run) %>% 
  summarise(t_test_p=t.test(fraction_1000x_coverage~sampletype)$p.value) %>% 
  arrange(desc(t_test_p)) %>% filter(t_test_p>0.01) %>% pull(run)
runs_to_drop2 = test_group %>% group_by(run) %>% 
  summarise(t_test_p=t.test(median_coverage~sampletype)$p.value) %>% 
  arrange(desc(t_test_p)) %>% filter(t_test_p>0.01) %>% pull(run)

############ 
runs_to_drop<-union(runs_to_drop1, runs_to_drop2)
saveRDS(runs_to_drop, "processing/runs_to_drop.rds")

mcov_samples_with_ct %>% pull(run) %>% unique()
test_group %>% group_by(run, sampletype) %>% summarize(counts = n()) %>% 
  filter(run %in% runs_to_drop) %>% nrow()

mcov_samples_with_ct %>% pull(run) %>% unique()
test_group %>% group_by(run, sampletype) %>% summarize(counts = n()) %>% 
  filter(sampletype=="negctrl") %>% arrange(counts) %>% mutate(dropped = run %in% runs_to_drop) %>%
  ggplot(aes(dropped,counts, label=run)) + geom_boxplot() + geom_point() + geom_text_repel() + theme_pubr()

runs_kept = mcov_samples_with_ct$run[!mcov_samples_with_ct$run %in% runs_to_drop] %>% unique()

```

```{r}
f1 = mcov_samples_with_ct %>% mutate(run_kept = run %in% runs_kept) %>% 
  ggplot(aes(INSTRUMENT_RESULT, fraction_1000x_coverage, 
                                         color = run_kept)) + 
  geom_point(shape=".", alpha = 0.5) + theme_pubr() + 
  theme(plot.title = element_text(size = 40, face = "bold")) + 
  geom_vline(xintercept=40, color = "red") + geom_smooth()
f1

f2 = mcov_samples_with_ct %>% mutate(run_kept = run %in% runs_kept) %>% 
  ggplot(aes(INSTRUMENT_RESULT, log10(median_coverage), color = run_kept)) + 
  geom_point(shape=".", alpha = 0.5) + theme_pubr() + 
  theme(plot.title = element_text(size = 40, face = "bold")) + 
  geom_vline(xintercept=40, color = "red") + geom_smooth()

runs_samples_dots = ggarrange(f1, f2, ncol = 1, nrow = 2, align = "v", common.legend = T)
runs_samples_dots
ggsave("ggsave/runs_samples_dots.pdf", plot = runs_samples_dots, width = 3, height = 4.5)
```


```{r}
mcov_samples_with_ct %>% colnames()
```


```{r}
### plot boxplot of the runs
stmp = mcov_samples_with_ct %>% mutate(run_dropped = run %in%   
    runs_to_drop) %>% mutate(run = as.numeric(gsub("Run_", "", run))) %>% 
    mutate(is_neg=if_else(CT>=40,1,0))

label_at <- function(n) function(x) ifelse(x %% n == 0, x, "")


f3 = ggplot(stmp, aes(x=run, y=fraction_1000x_coverage, group = run, 
             color = run_dropped, fill = run_dropped)) + 
  geom_boxplot() + 
  geom_point(data = stmp %>% filter(is_neg == T), 
             aes(run, fraction_1000x_coverage), color = "red", alpha = .9, 
             shape = 1) + 
  scale_fill_grey(start = 1, end = 0.8) + 
  scale_x_continuous(breaks = seq(10,90,2), limits = c(10,92), labels = label_at(10)) +  
  scale_color_grey(start=0, end=0.6) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5)) +
  labs( x = NULL)
  
f4 = ggplot(stmp, aes(x=run, y=log10(median_coverage), group = run, 
             color = run_dropped, fill = run_dropped)) + 
  geom_boxplot() + 
  geom_point(data = stmp %>% filter(is_neg == T), 
             aes(run, log10(median_coverage)), color = "red", alpha = .9, 
             shape = 1) + 
  scale_fill_grey(start = 1, end = 0.8) + 
  scale_x_continuous(breaks = seq(10,90,2), limits = c(10,92), labels = label_at(10)) +  
  scale_color_grey(start=0, end=0.6) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5)) +
  labs( x = NULL)

run_filter_red = ggarrange(f3, f4, ncol = 1, nrow = 2, align = "v", common.legend = T)
((
  run_filter_red_plot = annotate_figure(run_filter_red,
               top = text_grob("red circles: samples Ct>40", color = "red", size = 10))
))

ggsave("ggsave/run_filter_red_plot.pdf", plot = run_filter_red_plot, width = 6, height = 4)
```

## Sample-level QC
```{r}
#nextclade QC
nc<-fread("houston_nextclade.tsv", sep='\t', data.table = F) %>% 
  mutate(MCoVNumber=regmatches(seqName, regexpr("[M,R,S,O]CoV.[0-9]+", seqName)) %>% 
           str_remove("-") %>% str_remove("_")) %>% filter(!duplicated(MCoVNumber))
nextclade_bad_samples<-nc %>% filter(qc.overallStatus %in% c("bad")) %>% pull(MCoVNumber)
#drop bad runs and samples that don't pass pangolin QC or nextclade QC
mcov_samples_filtered<-mcov_samples %>% filter(!run %in% runs_to_drop) %>% 
  filter(qc_status=="pass") %>% filter(!MCoVNumber %in% nextclade_bad_samples) %>% 
  filter(scorpio_call!="Omicron (BA.1-like)") %>% 
 ##### #main coverage criterion for fair comparisons: X depth over Y percent of the genome
  filter(fraction_100x_coverage>=0.98) %>% droplevels()

#what's the distribution of CT values?
ct_distribution_after_qc<-mcov_samples_filtered %>% filter(INSTRUMENT_RESULT<50) %>% 
  ggplot(aes(x=INSTRUMENT_RESULT)) + geom_histogram(binwidth=1) + theme_bw() + 
  xlab("Ct value") + labs(caption="After exclusion criteria")
```

```{r}
#How did exclusion criteria change distribution of sample Ct values?

ct_distribution_after_qc<-mcov_samples_filtered %>% filter(INSTRUMENT_RESULT<50) %>% ggplot(aes(x=INSTRUMENT_RESULT)) + geom_histogram(binwidth=1) + theme_bw() + xlab("Ct value")

ct_distribution_before_qc<-mcov_samples %>% filter(INSTRUMENT_RESULT<50) %>% ggplot(aes(x=INSTRUMENT_RESULT)) + geom_histogram(binwidth=1) + theme_bw() + xlab("Ct value") + labs(caption="Before exclusion criteria")
#ct_inclusion<-plot_grid(ct_distribution_before_qc, ct_distribution_after_qc, nrow=2)

#ct_inclusion


ct_distribution_after_qc<-mcov_samples_filtered %>% filter(INSTRUMENT_RESULT<50)
ct_distribution_before_qc<-mcov_samples %>% filter(INSTRUMENT_RESULT<50)
plot1_ct_before_after = ggplot(aes(x=INSTRUMENT_RESULT), data = ct_distribution_before_qc) + 
  geom_histogram(binwidth = 1, fill = "grey") +
  geom_histogram(data = ct_distribution_after_qc, binwidth=1) + theme_pubr() + 
  xlab("Ct value") + ylab("# samples") #+ geom_vline(xintercept = 26,
                                               #     linetype = "dashed")

ggsave("ggsave/plot1_ct_before_after.pdf", plot1_ct_before_after, height = 2, width = 4)
```

## Preliminary minor variant distributions

```{r}
#nucleotide positions of all primers used in dataset; will exclude 
primers = fread("nCoV-2019.artic_v3.primer.txt", sep="\t", header=FALSE, data.table=F) %>% 
  select(start = V2, end = V3)
primer_positions_v3<-as.numeric()
for (i in 1:nrow(primers)){
  primer_positions_v3<-c(primer_positions_v3, primers[i,]$start:primers[i,]$end)
}

primers = fread("nCoV-2019.artic_v4.primer.bed", sep="\t", 
                header=FALSE, data.table = F) %>% select(start = V2, end = V3)
primer_positions_v4<-as.numeric()
for (i in 1:nrow(primers)){
  primer_positions_v4<-c(primer_positions_v4, primers[i,]$start:primers[i,]$end)
}

problem_sites_global = fread("problematic_sites_sarsCov2_v8-20211027.vcf", skip = 88) %>%
  filter(FILTER != "mask") %>%
  filter(!grepl('single_src|nanopore', INFO)) %>% pull(POS)
problem_sites_houston = fread("problematic_sites_sarsCov2_v8-20211027.vcf", skip = 88) %>%
  filter(grepl('Houston', INFO)) %>% pull(POS)
problem_sites = unique(c(problem_sites_global, problem_sites_houston))

primer_positions_all <- c(primer_positions_v3, primer_positions_v4, problem_sites) %>% unique() 
# insert problem sites if necessary

saveRDS(primer_positions_all, "primer_positions_all.rds")

#reference genome with nucleotide positions of genes
genes <- fread("ntpos_gene_update.csv", data.table = F)
gene_names <- genes %>% pull(gene_id) %>% unique()
genes$gene_id <- factor(genes$gene_id, levels = gene_names)
```

```{r}
### Update this if you change sample inclusion criteria
### Update this if change the thresholds for counting minor variants
#file was already filtered to sites with minimum 100 reads depth and A,C,T,G minor variant present and binomial significance check passed. Further filtering:


depth_at_site<-100
minor_frequency<-0.01
total_minor_reads<-50

minor_variant_sites_threshold <- fread("minor_sites_100x_all_20220507.csv", data.table = F) %>%
  mutate(MCoVNumber=regmatches(name,
                               regexpr("[M,R,S,O]CoV.[0-9]+", name)) %>%
           str_remove("-") %>% str_remove("_")) %>%
  filter(!ntpos %in% primer_positions_all) %>% #don't include primer binding sites
  filter(MCoVNumber %in% mcov_samples_filtered$MCoVNumber) %>%
  filter(totalcount>=depth_at_site) %>%
  filter(!ntpos %in% c(1:265)) %>% filter(!ntpos>29674) %>% #don't include 5' and 3' UTR
  filter(major %in% c("A","C","T","G")) %>% #don't want minor variants at consensus deletion sites
  filter(minorfreq>=minor_frequency) %>%
  filter(minorfreq*totalcount>=total_minor_reads)
write_feather(minor_variant_sites_threshold, 'processing/minor_variants_filtered_100x0.01_50.arrow')
```



## Overall minor variant richness

```{r}
minor_variant_sites_threshold <- read_feather('processing/minor_variants_filtered_100x0.01_50.arrow') 
n_var <- minor_variant_sites_threshold %>% group_by(MCoVNumber) %>% tally() %>% arrange(desc(n)) #this tally doesn't include any samples with 0 variants, so need to join to original list
samples_n_var <- mcov_samples_filtered %>% left_join(n_var) %>% 
  arrange(COLLECTION_DT) %>% mutate(n_var=tidyr::replace_na(n,0))
write_feather(samples_n_var, "processing/samples_n_var.arrow")

#what's the distribution of minor variant richness?
### 
((
  n_var_select <- samples_n_var %>% ggplot(aes(x=n_var)) + 
  geom_histogram(binwidth=5) + theme_pubr() + 
  xlab("No. minor variants in sample") + ylab("No. samples") + 
  scale_y_continuous(trans='log1p', breaks=c(1, 10, 100, 1000, 5000))
))
```

```{r}
#what's the relationship between minor variant richness and Ct value?


n_var_select_ct = samples_n_var %>% filter(INSTRUMENT_RESULT<50) %>% 
  ggplot(aes(x=INSTRUMENT_RESULT, y=n_var)) + 
  geom_point(shape = 1, alpha = 1/8) + 
  geom_density_2d(alpha = 1/2, color = "red") + 
  scale_y_continuous(trans = "log1p", breaks = c(0,1,5,10,30,50,100,300,500)) +
  scale_x_continuous(breaks = seq(0,40,by = 5)) +
  theme_pubr() + xlab("Ct value") + ylab("n_var") +
  geom_hline(yintercept = 30, linetype = "dashed") + 
  geom_vline(xintercept = 26, linetype = "dashed")

plot1_target = ggMarginal(n_var_select_ct, type = "violin", draw_quantiles = 
             c(.25,.5,.75))

ggsave("ggsave/plot1_target.pdf", plot = plot1_target, height = 3, width = 4)

((
n_var_select_ct_filtered = samples_n_var %>% 
  filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% 
  ggplot(aes(x=INSTRUMENT_RESULT, y=n_var)) + 
  geom_point(shape = 1, alpha = 1/16) + 
  geom_density_2d(color = "salmon") + 
  scale_y_continuous(breaks = seq(0,30,5)) +
  scale_x_continuous(breaks = seq(0,40,by = 5)) +
  theme_pubr() + xlab("Ct value") + ylab("n_var") +
  geom_smooth(color = "red") + stat_cor(method = "spearman", cor.coef.name = "rho")
))

patient_ordinal_counts = samples_n_var %>% 
  filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% 
  mutate(ordinal_counts = cut(n_var, breaks = seq(0,30,5), 
                              right = F)) %>% 
  select(n_var, ordinal_counts) %>% arrange(n_var) %>% 
  group_by(ordinal_counts) %>% summarize(counts = n()) %>% 
  mutate(position = seq(2.5,27.5,5))

n_var_select_ct_filtered_labeled = n_var_select_ct_filtered +
  geom_text(aes(x=25, y = position, label = counts), color = "darkblue", 
                                     data = patient_ordinal_counts) +
  theme_minimal_hgrid()

ct_n_var_postfilter_plot = ggMarginal(n_var_select_ct_filtered_labeled, 
                                      type = "violin", 
                                      draw_quantiles = c(0.25,0.5,0.75))
ct_n_var_postfilter_plot


ggsave("ggsave/ct_n_var_postfilter_plot.pdf", 
       ct_n_var_postfilter_plot, height = 3, width = 3)

saveRDS(ct_n_var_postfilter_plot, "ct_n_var_postfilter_plot.rds")
```

```{r}
# Quantile information for the violin plot above
print(samples_n_var %>% 
  filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% pull(n_var) %>% quantile(.,probs = c(0,0.25,0.5,0.75)))

print(samples_n_var %>% 
  filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% pull(INSTRUMENT_RESULT) %>% quantile(.,probs = c(0,0.25,0.5,0.75)))
```

```{r}
samples_n_var %>% pull(n_var) %>% summary()
```

## Reproducibility of minor variants


```{r}
minors_table<- fread("replicated_samples.csv", data.table=F) %>% 
  filter(MCoVNumber %in% samples_n_var$MCoVNumber) %>% 
  filter(major.original %in% c("A","C","T","G")) %>% 
  filter(minor.original %in% c("A","C","T","G")) %>%
  filter(totalcount.original>=100 & minorfreq.original>=0.01 & totalcount.original*minorfreq.original>=50 & 
           tolower(binocheck.original)!="false") %>% filter((!ntpos %in% primer_positions_all)) %>% 
  filter(!ntpos %in% 1:265) %>% filter(!ntpos>29674)
#was the same minor variant found in the second replicate? if not, set minor frequency to 0 in second rep
minors_table<-minors_table %>% 
  mutate(minorfreq.reseq=if_else(minor.original==minor.reseq, minorfreq.reseq, 0)) %>% 
  mutate(detected_minor_in_repl=if_else(minor.original==minor.reseq,"yes","no"))

write_feather(minors_table, "processing/minors_table.arrow")
```

```{r}
#how well are minor variants recovered in samples with different Ct values?
minors_table <- minors_table %>% left_join(select(mcov_samples, MCoVNumber, CT=INSTRUMENT_RESULT)) %>% 
  mutate(sampleCT_bin = case_when(CT<26 ~ "below 26",
                                CT>=26&CT<=35 ~"CT 26-35",
                                CT>35&CT<50 ~"greater than 35",
                                CT>100~"unknown", #aptima instrument uses RLU not CT
                                is.na(CT) ~ "unknown")) 
((
  rep_ct<-minors_table %>% 
  ggplot(aes(x=minorfreq.original, y=minorfreq.reseq, color=detected_minor_in_repl)) + 
  scale_color_manual(values=c("red","black")) + geom_point(alpha=0.5) + 
  facet_wrap(~sampleCT_bin) + theme_bw() + labs(x="MAF in replicate 1", y="MAF in replicate 2") + 
  theme(legend.position="bottom")
))

```

```{r}
#how well are minor variants recovered in samples with different median coverage?
rep_depth <- minors_table %>% left_join(coverage_levels, by=c("MCoVNumber"="samplename")) %>% 
  mutate(coverage_bin=cut(median_coverage, 4)) %>% 
  ggplot(aes(x=minorfreq.original, y=minorfreq.reseq, color=detected_minor_in_repl)) + 
  scale_color_manual(values=c("red","black")) + geom_point(alpha=0.5) + 
  facet_wrap(~coverage_bin) + theme_bw() + 
  labs(x="MAF in replicate 1", y="MAF in replicate 2") + 
  theme(legend.position="none", axis.text.x = element_text(color=c(1,0,1,0))) + 
  labs(caption="Reproducibility in samples with different median depths")
#in the range of Ct values/coverage observed here, reproducibility seems more associated with Ct than with coverage 
```

```{r}
#what's the distribution of depth/MAF in reproducible vs. non-reproducible minor variants?
rep_depth_freq <- minors_table %>% ggplot(aes(x=totalcount.original, y=minorfreq.original)) + 
  geom_point(alpha=0.5) + facet_grid(detected_minor_in_repl~.) + theme_bw() + 
  theme(axis.text.x = element_text(color=c(1,0,1,0))) + xlab("Seq depth in rep 1") + 
  ylab("MAF in rep 1") + labs(caption="Reproducibility at sites with different depth and MAF")
#depth and frequency of minor variants is also not very different between reproducible and non-reproducible variants
```

```{r}
rep_mutations <- minors_table %>% filter(detected_minor_in_repl=="yes") 
nonrep_mutations <- minors_table %>% filter(detected_minor_in_repl=="no") 

plot_reproducible_spectra_heatmap <- table(rep_mutations$major.original, rep_mutations$minor.original) %>% 
  data.frame() %>% ggplot(aes(x=Var1, y=Var2, fill=Freq/sum(Freq))) + 
  geom_tile(colour = "black") + # grid colour
  scale_fill_gradient(low = "white",
                      high = "darkblue") +
  theme_minimal() + labs(fill = "Fraction",
       x = "Consensus allele",
       y = "Minor allele", title="Reproducible")


plot_nonreproducible_spectra_heatmap <- table(nonrep_mutations$major.original,
                                              nonrep_mutations$minor.original) %>% 
  data.frame() %>% ggplot(aes(x=Var1, y=Var2, fill=Freq/sum(Freq))) + 
  geom_tile(colour = "black") + 
  scale_fill_gradient(low = "white",
                      high = "darkblue") +
  theme_minimal() + labs(fill = "Fraction",
       x = "Consensus allele",
       y = "Minor allele", title="Non-reproducible")

saveRDS(plot_nonreproducible_spectra_heatmap, file = "plot_nonreproducible_spectra_heatmap.rds")
saveRDS(plot_reproducible_spectra_heatmap, file = "plot_reproducible_spectra_heatmap.rds")

# Plotted in RMD 03.
```


```{r}
rep_indiv_samples <- minors_table %>% ggplot(aes(x=minorfreq.original, 
                                               y=minorfreq.reseq, 
                                               color=detected_minor_in_repl)) + 
  geom_point(alpha=0.5) + scale_color_manual(values=c("red","black")) + 
  facet_wrap(CT~MCoVNumber) + theme_bw() + xlab("MAF in replicate 1") + 
  ylab("MAF in replicate 2") + 
  theme(legend.position="none", axis.text.x = element_text(color=c(1,0,1,0)))
```

```{r, fig.height = 10}
#SUPP FIG 1
replicate_mega_plot = plot_grid(
  plot_grid(rep_ct, 
            plot_grid(rep_depth, rep_depth_freq, ncol=2, rel_widths=c(1.2,1), 
                      labels=c("b","c")), nrow=2, labels=c("a",NA)), 
  rep_indiv_samples, labels=c(NA, "d"), rel_widths=c(1.1,1))

ggsave("ggsave/replicate_mega_plot.pdf", plot = replicate_mega_plot, height = 8, width = 14.5)

```


Will limit analyses to samples with CT<26, where we are more confident in reproducibility of minor variant


```{r}
samples_n_var %>% filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% pull(n_var) %>% quantile(c(0.5,0.7))
```

```{r}
samples_to_analyze <- samples_n_var %>% filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% pull(MCoVNumber)
lineages_figure <- mcov_samples_filtered %>% 
  filter(MCoVNumber %in% samples_to_analyze) %>% 
  mutate(variant=case_when(startsWith(scorpio_call,"Delta") ~ "Delta",
        startsWith(scorpio_call,"Alpha") ~ "Alpha",
        !(startsWith(scorpio_call,"Delta")|
            startsWith(scorpio_call,"Alpha")) ~ "other")) %>%
  mutate(variant=if_else(lineage == "B.1.2","B.1.2",variant)) %>%
  mutate(month = floor_date(COLLECTION_DT, "month")) %>%
  ggplot(aes(x=month)) + geom_bar(aes(fill=variant))  + 
#  scale_fill_manual(values=c("#00A08A", "#F2AD00", "#F98400", "#5BBCD6")) + 
  xlab("Collection date") + ylab("No. samples") +   scale_x_date(labels = date_format("%m-%Y")) +
  ggtitle(paste0('Final sample set, \n Ct<26 & n_var<30, n = ',length(samples_to_analyze))) + 
  scale_x_date(minor_breaks="1 month") + theme_pubr() +
  theme(axis.text.x = element_text(angle = 90, vjust = .5), legend.position = "top")

#SUPP FIG 2
lineages_figure
saveRDS(lineages_figure, "ggsave/lineages_figure.rds")
```

```{r}
samples_n_var %>% filter(INSTRUMENT_RESULT<26) %>% pull(n_var) %>% summary()
```

```{r}
#plot_grid(coverage_pre_filtering, ct_inclusion, ncol=2, labels=c("a","b"))
```

```{r}
minor_sites_lowct<-minor_variant_sites_threshold %>% left_join(mcov_samples_filtered) %>% 
  filter(INSTRUMENT_RESULT<26) 
```


# What kinds of run-specific effects do we see even after filtering for high-quality samples?

```{r}
#are run effects related to sequencing depth?
n_var_by_coverage<-samples_n_var %>% filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% ggplot(aes(x=median_coverage, y=n_var)) + geom_point(alpha=0.05) + 
  geom_density_2d(color = "salmon") + geom_smooth(color = "red") + theme_pubr() +
  scale_x_continuous(limits = c(0, 15000)) +
  xlab("Sample median coverage") + ylab("n_var") + stat_cor(method = "spearman", cor.coef.name = "rho")
n_var_by_coverage
#not on the individual sample level

#is average minor variant richness in a run related to average depth of coverage in the run?
n_var_depth_averages<-samples_n_var %>% filter(INSTRUMENT_RESULT<26 & n_var < 30) %>% 
  group_by(run) %>% 
  summarise(median_n_var=median(n_var), 
            median_sample_coverage=median(median_coverage), 
            median_ct=median(INSTRUMENT_RESULT)) %>% 
  ggplot(aes(x=median_sample_coverage, y=median_n_var)) + geom_point(alpha = 0.5) +
  scale_x_continuous(limits = c(0, 15000)) +
  theme_pubr() + xlab("Run median of median coverage") + ylab("Run median n_var") + 
  geom_smooth(color = "red") + stat_cor(method = "spearman", cor.coef.name = "rho")
n_var_depth_averages
```

```{r, fig.height = 5, fig.width = 4}
n_var_by_run <- samples_n_var %>% mutate(run = as.numeric(str_remove(run, "Run_"))) %>% 
  filter(INSTRUMENT_RESULT < 26 & n_var < 30) %>% 
    ggplot(aes(x=run, y=n_var, group = run)) + geom_boxplot() + 
    scale_fill_grey(start = 1, end = 0.8) + 
    scale_x_continuous(breaks = seq(10,90,2), limits = c(10,92), labels = label_at(10)) +
    scale_y_continuous(breaks = seq(0,30,2), labels = label_at(10)) +

    scale_color_grey(start=0, end=0.6) + 
    theme(axis.text.x = element_text(angle = 90, vjust = .5))

plots_by_run = ggarrange(f3, f4, n_var_by_run, labels = list("A","B","D"), 
          ncol = 1, nrow = 3, align = "v", common.legend = T) %>% annotate_figure(.,
               top = text_grob("red circles: samples Ct>40", color = "red", size = 10))

dots_depth_median = ggarrange(n_var_by_coverage, n_var_depth_averages,  
                              labels = list("E", "F"), ncol = 1)

dots_plots = ggarrange(runs_samples_dots, dots_depth_median, labels = list("C"), 
                       ncol = 1, heights = c(1,1))

run_filtration = ggarrange(plots_by_run, dots_plots, ncol = 2, widths = c(2.5, 1))
run_filtration
ggsave("ggsave/run_filtration.pdf", run_filtration, height = 8, width = 12)
```
